# docs/swagger.yaml
basePath: /v1/core-services
definitions:
...
  github_com_modami_core-service_internal_dto.CreateProductRequest:
    properties:
      brand:
        type: string
      category_id:
        type: string
      color:
        type: string
      condition:
        enum:
        - new
        - like_new
        - good
        - fair
        type: string
      credit_cost:
        type: integer
      description:
        maxLength: 5000
        minLength: 20
        type: string
      hashtags:
        items:
          type: string
        maxItems: 10
        type: array
      images:
        items:
          $ref: '#/definitions/github_com_modami_core-service_internal_dto.ImageInput'
        maxItems: 6
        minItems: 1
        type: array
      material:
        type: string
      price:
        minimum: 1000
        type: integer
      size:
        type: string
      title:
        maxLength: 200
        minLength: 5
        type: string
    required:
    - category_id
    - condition
    - description
    - images
    - price
    - size
    - title
    type: object
...
  github_com_modami_core-service_internal_dto.ImageInput:
    properties:
      height:
        type: integer
      position:
        type: integer
      url:
        type: string
      width:
        type: integer
    required:
    - url
    type: object
  github_com_modami_core-service_internal_dto.ResubmitRequest:
    properties:
      brand:
        type: string
      category_id:
        type: string
      color:
        type: string
      condition:
        enum:
        - new
        - like_new
        - good
        - fair
        type: string
      description:
        maxLength: 5000
        minLength: 20
        type: string
      hashtags:
        items:
          type: string
        maxItems: 10
        type: array
      images:
        items:
          $ref: '#/definitions/github_com_modami_core-service_internal_dto.ImageInput'
        maxItems: 6
        minItems: 1
        type: array
      material:
        type: string
      note:
        type: string
      price:
        minimum: 1000
        type: integer
      size:
        type: string
      title:
        maxLength: 200
        minLength: 5
        type: string
    type: object
...
  github_com_modami_core-service_internal_dto.UpdateProductRequest:
    properties:
      brand:
        type: string
      category_id:
        type: string
      color:
        type: string
      condition:
        enum:
        - new
        - like_new
        - good
        - fair
        type: string
      credit_cost:
        type: integer
      description:
        maxLength: 5000
        minLength: 20
        type: string
      hashtags:
        items:
          type: string
        maxItems: 10
        type: array
      images:
        items:
          $ref: '#/definitions/github_com_modami_core-service_internal_dto.ImageInput'
        maxItems: 6
        minItems: 1
        type: array
      material:
        type: string
      price:
        minimum: 1000
        type: integer
      size:
        type: string
      title:
        maxLength: 200
        minLength: 5
        type: string
    type: object
...
  internal_adapter_handler.StandardErrorEnvelope:
    properties:
      error: {}
      meta: {}
      success:
        type: boolean
    type: object
  internal_adapter_handler.StandardSuccessEnvelope:
    properties:
      data: {}
      meta: {}
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'Core service: catalog, orders, master data, and seller APIs for the
    Modami marketplace.'
  title: Modami Core Service API
  version: "1.0"
paths:
...
  /hashtags/{tag}/products:
    get:
      parameters:
      - description: 'Hashtag (without #)'
        in: path
        name: tag
        required: true
        type: string
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Products by hashtag
      tags:
      - Products
...
  /products:
    post:
      consumes:
      - application/json
      parameters:
      - description: Product payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_modami_core-service_internal_dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create product
      tags:
      - Products
  /products/{id}:
    delete:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - Products
    get:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Get product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_modami_core-service_internal_dto.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - Products
  /products/{id}/archive:
    post:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Archive product
      tags:
      - Products
  /products/{id}/moderation:
    get:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: List moderation history for product
      tags:
      - Products
  /products/{id}/resubmit:
    post:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Updates
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_modami_core-service_internal_dto.ResubmitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Resubmit rejected product
      tags:
      - Products
  /products/{id}/similar:
    get:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Max items (default 10, max 50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Similar products
      tags:
      - Products
  /products/{id}/submit:
    post:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Submit product for review
      tags:
      - Products
  /products/{id}/unarchive:
    post:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Unarchive product
      tags:
      - Products
  /products/{id}/view:
    post:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Track product view
      tags:
      - Products
  /products/featured:
    get:
      parameters:
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Featured products
      tags:
      - Products
  /products/feed:
    get:
      parameters:
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Product feed (cursor)
      tags:
      - Products
  /products/me:
    get:
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List my products (seller)
      tags:
      - Products
  /products/search:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Category filter
        in: query
        name: category_id
        type: string
      - description: Condition
        in: query
        name: condition
        type: string
      - description: Brand
        in: query
        name: brand
        type: string
      - description: Min price
        in: query
        name: min_price
        type: integer
      - description: Max price
        in: query
        name: max_price
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Search products (catalog)
      tags:
      - Search
  /products/select:
    get:
      parameters:
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Curated select products
      tags:
      - Products
  /products/slug/{slug}:
    get:
      parameters:
      - description: URL slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Get product by slug
      tags:
      - Products
...
  /sellers/{id}/products:
    get:
      parameters:
      - description: Seller user ID
        in: path
        name: id
        required: true
        type: string
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: Seller's listed products
      tags:
      - Sellers
...
  /sellers/{seller_id}/products:
    get:
      parameters:
      - description: Seller user ID
        in: path
        name: seller_id
        required: true
        type: string
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      summary: List products by seller
      tags:
      - Products
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"